scrapers/pyproject.toml
[project]
name = "cardpulse-scrapers"
version = "0.1.0"
description = "Data scrapers for CardPulse - eBay sales, PSA population, and live listings"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
